file cleanup changes
